tests/requirements.txt
graphql-core==3.2.3
attrs==22.1.0
bson==0.5.10
docstring-parser==0.15
pydantic==1.10.2
pytest==7.2.0
pytest-cov==4.0.0
pytest-asyncio==0.20.3
SQLAlchemy==1.4.45
typing-extensions==4.4.0
jsonschema==4.17.3
